requirements.txt
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.11
fonttools==4.53.1
importlib_resources==6.4.5
kiwisolver==1.4.7
matplotlib==3.7.5
numpy==1.24.4
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pycocotools==2.0.7
pyparsing==3.1.4
python-dateutil==2.9.0.post0
scipy==1.10.1
six==1.16.0
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111
tqdm==4.66.5
typing_extensions==4.12.2
zipp==3.20.2
